requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
psycopg2-binary==2.9.9
langchain-groq==0.1.6
langchain-community==0.2.6
langchain-core==0.2.11
duckduckgo-search==6.1.9
pydantic==2.7.4
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==23.2.1
requests
beautifulsoup4
lxml
httpx
tavily-python
